% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pros.mlops.1-package.R
\docType{package}
\name{pros.mlops.1-package}
\alias{pros.mlops.1}
\alias{pros.mlops.1-package}
\title{pros.mlops.1: What the Package Does (Predict, Pros)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
